Split `TwoPhaseValidation`into 5 modules
