+++ libs/cardano-ledger-test/test/Tests.hs
@@ -11,12 +11,12 @@ import Data.Default.Class (Default (def))
 import System.IO (hSetEncoding, stdout, utf8)
 import qualified Test.Cardano.Ledger.Alonzo.Tools as Tools
 import Test.Cardano.Ledger.BaseTypes (baseTypesTests)
+import qualified Test.Cardano.Ledger.Examples.AlonzoAPI as AlonzoAPI (tests)
+import qualified Test.Cardano.Ledger.Examples.AlonzoBBODY as AlonzoBBODY (tests)
+import qualified Test.Cardano.Ledger.Examples.AlonzoCollectInputs as AlonzoCollectInputs (tests)
+import qualified Test.Cardano.Ledger.Examples.AlonzoInvalidTxUTXOW as AlonzoInvalidTxUTXOW (tests)
+import qualified Test.Cardano.Ledger.Examples.AlonzoValidTxUTXOW as AlonzoValidTxUTXOW (tests)
 import Test.Cardano.Ledger.Examples.BabbageFeatures (babbageFeatures)
-import Test.Cardano.Ledger.Examples.TwoPhaseValidation
-  ( allTrees,
-    alonzoAPITests,
-    collectOrderingAlonzo,
-  )
 import Test.Cardano.Ledger.Generic.AggPropTests (aggTests)
 import Test.Cardano.Ledger.Generic.Properties (genericProperties)
 import Test.Cardano.Ledger.Model.Properties (modelUnitTests_)
@@ -38,10 +38,12 @@ mainTestTrees =
     Tools.tests,
     testGroup
       "STS Tests"
-      [ allTrees,
-        babbageFeatures,
-        alonzoAPITests,
-        collectOrderingAlonzo,
+      [ babbageFeatures,
+        AlonzoValidTxUTXOW.tests,
+        AlonzoInvalidTxUTXOW.tests,
+        AlonzoBBODY.tests,
+        AlonzoAPI.tests,
+        AlonzoCollectInputs.tests,
         modelUnitTests_
       ],
     genericProperties def,
